Capstone_Winter_Work/FrameworkInstanceExamples/HVACZigbee/ventDriver/requirements.txt
chardet==2.3.0
cryptography==1.7.1
idna==2.2
keyring==10.1
keyrings.alt==1.3
netifaces==0.10.9
pyasn1==0.1.9
pycrypto==2.6.1
pygobject==3.22.0
python-apt==1.1.0b5
pyxdg==0.25
requests==2.12.4
RPi.GPIO==0.6.5
SecretStorage==2.3.1
six==1.10.0
ssh-import-id==5.6
urllib3==1.19.1
